refactor setNodeValues action
